+++ components/country/CountryUserTable.js
@@ -1,40 +1,24 @@
 import React from 'react'
-import Link from '../Link'
-import { sortBy, prop } from 'ramda'
-import { formatDecimal } from '../../lib/utils/format'
-import TableHeaders from '../common/TableHeaders'
-import { tableHeaderNames } from '../../lib/enums'
+import Table from '../common/Table'
+
+const tableSchema = {
+  'headers': {
+    'name': { type: 'namelink', accessor: 'full_name' },
+    'edits': { type: 'number', accessor: 'edits' },
+    'changesets': { type: 'number', accessor: 'changesets' }
+  },
+  'columnOrder': ['name', 'edits', 'changesets'],
+  'displaysTooltip': [
+    'edits',
+    'changesets'
+  ]
+}
 
 export default function CountryUserTable (props) {
+  const idMap = Object.assign(...props.users.map(({ osm_id, full_name }) => ({ [full_name]: osm_id })))
   return (
     <div className='widget'>
-      <table>
-        <thead>
-          <tr>
-            <TableHeaders tableName={tableHeaderNames.COUNTRY_USER} />
-          </tr>
-        </thead>
-        <tbody>
-          {
-            sortBy(prop('edits'), props.users)
-              .reverse()
-              .map((user, idx) => (
-                <tr key={user.osm_id}>
-                  <td>{idx + 1}</td>
-                  <td>
-                    <Link href={`/users/${user.osm_id}`}>
-                      <a className='link--normal' >
-                        {user.full_name}
-                      </a>
-                    </Link>
-                  </td>
-                  <td>{formatDecimal(user.edits)}</td>
-                  <td>{formatDecimal(user.changesets)}</td>
-                </tr>
-              ))
-          }
-        </tbody>
-      </table>
+      <Table tableSchema={tableSchema} idMap={idMap} data={props.users} initialSortColumn='edits' />
     </div>
   )
 }