+++ pages/_app.js
@@ -194,7 +194,7 @@ class Layout extends React.Component {
                           <ul>
                             <li><Link href='/dashboard'><a>Dashboard</a></Link></li>
                             <li><Link href={`/users/${osm.id}`}><a>Public Profile</a></Link></li>
-                            <li><Link href={`/edit-user?id=${osm.id}`} as={join(APP_URL_PREFIX, `/users/${osm.id}/edit`)}><a>Edit Profile</a></Link></li>
+                            <li><Link href={`/edit-user?id=${osm.id}`} as={`/users/${osm.id}/edit`}><a>Edit Profile</a></Link></li>
                             {
                               account.roles && isAdmin(account.roles) && (
                                 <li><Link href={`/admin`}><a>Admin</a></Link></li>