+++ CHANGELOG.md
@@ -6,9 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+### Changed
+- Refactor tables and tableheaders to use a common component configured using a tableSchema
+- Make all tables sortable
+
 ### Fixed
 - Removed "You" from public profile sidebars
 - Design fixes: ensured full-height application, aligned country maps, included full height headers and loading states
+- Removed source environment URL join from Edit Profile link in menu
+- Replace `edit_sum` on User pages with the updated `edit_count` from OSMesa
+- Fixes bug in tables when defaults for statistics are not provided by the API
 
 ## [v1.6.0] - 2019-10-28
 ### Changed